[WIP] Deriving parents, take 2
There is a temporary command, "find-branches", which will scan over the
changelog and try to write out change_parents records.

The logic for what is considered a cross-merge is still taking shape.
After this, there will be the job of calculating what integration records
contradict the change_parents, and reporting on them in the commit message.