Approaching the 99% good enough point
The merge detection is now very good - good enough for it to run over much
of the Perl history without much interaction.  Turns out that *all* of the
integration history from the merge base to the merged heads must be
examined.  It's a bit of a hack - one of the most expensive queries that
we do could probably still be eliminated.  Time to start exporting some
commits.