Load p4raw-extra-\w+.asv files straight after checkpoints
The 'p4raw-extra-users.asv' files, etc, fill in extra rows that are
required for the constraints checks to pass.  As the journal files contain
updates and deletes, they can't be loaded during the initial pass.  So,
put these filler files in the load order first.

This makes use of a distinction between the filenames which is a bit of a
hack - that 'p4raw-extra-\w+.asv' is intended to be loaded BEFORE
constraints are added, and 'p4raw-extra-journal-\d+.asv.gz' is loaded
AFTER.