Handle @rv@ and @dv@
I'm really quite happy with this.  I ended up re-using a code path that I
had previously added just to make the loads faster!  Basically when it
decides that we're loading a journal file, it will load to a temporary
table - taking the branch previously selected manually with the --reload
option.

Once all the rows are loaded, it uses DELETE ... USING to remove the old
versions of the rows from the target table.  Then, it uses a
SELECT DISTINCT ON ... to load the last inserted values of all rows to the
target tables.  For this to work, all FK constraints need to be deferred.

There is a minor overhead here to add an integer column to every table,
that records which source file the row was deleted in.  This is modeled
loosely after the xmax system Pg column.  A subselect ensures that rows
which have a maximum set are not actually loaded into the main tables.