Attempt to fix an escape decoding bug
The dump format delimits fields with @ signs, and uses a double-@ inside
the string to escape literal @'s.  In the case that a @-quoted string has
embedded newlines, the code reads extra lines until it sees the end of the
string.  However, this assumed that the multi-line @@ string is at the end
of the row.  Allow it to be a non-terminal field.